+++ config/ui.php
@@ -1,0 +1,29 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+
+    'prefix' => 'vee',
+
+    'theme' => [
+        'colors' => [
+            'primary' => 'indigo',
+            'secondary' => 'violet',
+            'accent' => 'lime',
+            'success' => 'green',
+            'info' => 'blue',
+            'warning' => 'yellow',
+            'error' => 'red',
+            'neutral' => 'zinc',
+        ],
+        'icons' => [
+            'dark' => 'lucide:moon',
+            'light' => 'lucide:sun',
+            'loading' => 'lucide:refresh-cw',
+            'system' => 'lucide:monitor',
+        ],
+        'transitions' => true,
+    ],
+
+];