+++ js/index.mjs
@@ -1,3 +1,0 @@
-export function isTrue(n) {
-    return n ? true : false;
-}