+++ packages/alpine/src/index.ts
@@ -1,0 +1,3 @@
+export function isTrue(n: string) {
+    return typeof n === 'string';
+}