+++ packages/vue/nuxt.config.ts
@@ -1,0 +1,4 @@
+export default defineNuxtConfig({
+    modules: ['@nuxt/eslint', '@nuxt/image', './src/module'],
+    eslint: { config: { standalone: false } },
+});