+++ src/Ui.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Vee;
 
 use Illuminate\Support\Facades\Facade;
+use Illuminate\Support\Facades\Config;
 
 /**
  * @see UiManager
@@ -13,6 +14,6 @@ final class Ui extends Facade
 {
     public static function getFacadeAccessor()
     {
-        return 'vee';
+        return Config::get('ui.prefix', 'vee');
     }
 }