+++ src/UiServiceProvider.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Vee;
 
 use Illuminate\Support\Facades\Blade;
+use Illuminate\Support\Facades\Config;
 use Illuminate\Support\ServiceProvider;
 use Illuminate\View\ComponentAttributeBag;
 use Illuminate\View\Factory as ViewFactory;
@@ -13,15 +14,19 @@ final class UiServiceProvider extends ServiceProvider
 {
     public function register(): void
     {
-        $this->app->alias(UiManager::class, 'vee');
+        $uiPrefix = Config::get('ui.prefix', 'vee');
+
+        $this->app->alias(UiManager::class, $uiPrefix);
         $this->app->singleton(UiManager::class);
 
         $loader = \Illuminate\Foundation\AliasLoader::getInstance();
-        $loader->alias('vee', Ui::class);
+        $loader->alias($uiPrefix, Ui::class);
     }
 
     public function boot(): void
     {
+        $uiPrefix = Config::get('ui.prefix', 'vee');
+
         $this->bootComponentPath();
         $this->bootTagCompiler();
         $this->bootMacros();
@@ -29,28 +34,32 @@ final class UiServiceProvider extends ServiceProvider
         app('livewire')->propertySynthesizer(DateRangeSynth::class);
         AssetManager::boot();
 
-        app('vee')->boot();
+        app($uiPrefix)->boot();
         $this->bootCommands();
     }
 
     public function bootComponentPath()
     {
+        $uiPrefix = Config::get('ui.prefix', 'vee');
+
         if (file_exists(resource_path('views/ui'))) {
-            Blade::anonymousComponentPath(resource_path('views/ui'), 'vee');
+            Blade::anonymousComponentPath(resource_path('views/ui'),  $uiPrefix);
         }
 
-        Blade::anonymousComponentPath(__DIR__.'/../stubs/resources/views/ui', 'vee');
+        Blade::anonymousComponentPath(__DIR__.'/../stubs/resources/views/ui',  $uiPrefix);
     }
 
     public function bootTagCompiler()
     {
+        $uiPrefix = Config::get('ui.prefix', 'vee');
+
         $compiler = new UiTagCompiler(
             app('blade.compiler')->getClassComponentAliases(),
             app('blade.compiler')->getClassComponentNamespaces(),
             app('blade.compiler')
         );
 
-        app()->bind('vee.compiler', fn () => $compiler);
+        app()->bind("{$uiPrefix}.compiler", fn () => $compiler);
 
         app('blade.compiler')->precompiler(function ($in) use ($compiler) {
             return $compiler->compile($in);