+++ src/UiTagCompiler.php
@@ -4,22 +4,40 @@ declare(strict_types=1);
 
 namespace Vee;
 
+use Illuminate\Support\Facades\Config;
 use Illuminate\View\Compilers\ComponentTagCompiler;
 use InvalidArgumentException;
 
+
 final class UiTagCompiler extends ComponentTagCompiler
 {
+
+    protected ?string $uiPrefix = null;
+
+    /**
+     * Get the UI prefix, fetching it from config if not already set.
+     */
+    protected function getUiPrefix(): string
+    {
+        if (is_null($this->uiPrefix)) {
+            $this->uiPrefix = Config::get('ui.prefix', 'vee');
+        }
+        return $this->uiPrefix;
+    }
+
     public function componentString(string $component, array $attributes)
     {
+        $uiPrefix = $this->getUiPrefix();
+
         // A component that forwards all data, attributes, and named slots to another component...
-        if ($component === 'vee::delegate-component') {
+        if ($component === "{$uiPrefix}::delegate-component") {
             $component = $attributes['component'];
 
             $class = \Illuminate\View\AnonymousComponent::class;
 
             // Laravel 12+ uses xxh128 hashing for views https://github.com/laravel/framework/pull/52301...
-            return "<?php if (!Ui::componentExists(\$name = {$component})) throw new \Exception(\"Ui component [{\$name}] does not exist.\"); ?>##BEGIN-COMPONENT-CLASS##@component('{$class}', 'vee::' . {$component}, [
-    'view' => (app()->version() >= 12 ? hash('xxh128', 'vee') : md5('vee')) . '::' . {$component},
+            return "<?php if (!Ui::componentExists(\$name = {$component})) throw new \Exception(\"Ui component [{\$name}] does not exist.\"); ?>##BEGIN-COMPONENT-CLASS##@component('{$class}', '{$uiPrefix}::' . {$component}, [
+    'view' => (app()->version() >= 12 ? hash('xxh128', '{$uiPrefix}') : md5('{$uiPrefix}')) . '::' . {$component},
     'data' => \$__env->getCurrentComponentData(),
 ])
 <?php \$component->withAttributes(\$attributes->getAttributes()); ?>";
@@ -37,10 +55,13 @@ final class UiTagCompiler extends ComponentTagCompiler
      */
     protected function compileOpeningTags(string $value)
     {
+        $uiPrefix = $this->getUiPrefix(); // Get prefix for regex
+        $escapedUiPrefix = preg_quote($uiPrefix, '/');
+
         $pattern = "/
             <
                 \s*
-                vee[\:]([\w\-\:\.]*)
+                {$escapedUiPrefix}[\:]([\w\-\:\.]*)
                 (?<attributes>
                     (?:
                         \s+
@@ -82,12 +103,12 @@ final class UiTagCompiler extends ComponentTagCompiler
             >
         /x";
 
-        return preg_replace_callback($pattern, function (array $matches) {
+        return preg_replace_callback($pattern, function (array $matches) use ($uiPrefix) {
             $this->boundAttributes = [];
 
             $attributes = $this->getAttributesFromAttributeString($matches['attributes']);
 
-            return $this->componentString("vee::{$matches[1]}", $attributes);
+            return $this->componentString("{$uiPrefix}::{$matches[1]}", $attributes);
         }, $value);
     }
 
@@ -100,10 +121,13 @@ final class UiTagCompiler extends ComponentTagCompiler
      */
     protected function compileSelfClosingTags(string $value)
     {
+        $uiPrefix = $this->getUiPrefix(); // Get prefix for regex
+        $escapedUiPrefix = preg_quote($uiPrefix, '/');
+
         $pattern = "/
             <
                 \s*
-                vee[\:]([\w\-\:\.]*)
+                {$escapedUiPrefix}[\:]([\w\-\:\.]*)
                 \s*
                 (?<attributes>
                     (?:
@@ -145,7 +169,7 @@ final class UiTagCompiler extends ComponentTagCompiler
             \/>
         /x";
 
-        return preg_replace_callback($pattern, function (array $matches) {
+        return preg_replace_callback($pattern, function (array $matches) use ($uiPrefix) {
             $this->boundAttributes = [];
 
             $attributes = $this->getAttributesFromAttributeString($matches['attributes']);
@@ -156,10 +180,10 @@ final class UiTagCompiler extends ComponentTagCompiler
 
                 unset($attributes['slot']);
 
-                return '@slot('.$slot.') '.$this->componentString("vee::{$matches[1]}", $attributes)."\n@endComponentClass##END-COMPONENT-CLASS##".' @endslot';
+                return '@slot('.$slot.') '.$this->componentString("{$uiPrefix}::{$matches[1]}", $attributes)."\n@endComponentClass##END-COMPONENT-CLASS##".' @endslot';
             }
 
-            return $this->componentString("vee::{$matches[1]}", $attributes)."\n@endComponentClass##END-COMPONENT-CLASS##";
+            return $this->componentString("{$uiPrefix}::{$matches[1]}", $attributes)."\n@endComponentClass##END-COMPONENT-CLASS##";
         }, $value);
     }
 
@@ -170,6 +194,9 @@ final class UiTagCompiler extends ComponentTagCompiler
      */
     protected function compileClosingTags(string $value)
     {
-        return preg_replace("/<\/\s*vee[\:][\w\-\:\.]*\s*>/", ' @endComponentClass##END-COMPONENT-CLASS##', $value);
+        $uiPrefix = $this->getUiPrefix();
+        $escapedUiPrefix = preg_quote($uiPrefix, '/');
+
+        return preg_replace("/<\/\s*{$escapedUiPrefix}[\:][\w\-\:\.]*\s*>/", ' @endComponentClass##END-COMPONENT-CLASS##', $value);
     }
 }