+++ turbo/tools/prettier/src/vue.mjs
@@ -1,0 +1,6 @@
+import baseConfig from './base.mjs';
+
+/** @type {import("prettier").Config} */
+export default {
+    ...baseConfig,
+};