+++ CHANGELOG.md
@@ -5,11 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.3]
+
+### Added
+
+- Add `examples` for Livewire
+- Add `examples` for Vue
+
 ## [0.0.2]
 
 ### Added
 
-- Add `ui` components (container, icon, link, main)
+- Add `ui` views (container, main)
+- Add `ui` views (icon, link)
 
 ## [0.0.1]
 
@@ -18,5 +26,5 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Add `ci` workflow (lint, test)
 - Add `test` workflow (pest, phpstan)
 - Add `lint` workflow (pint, prettier)
-- Add `build` workflow (esbuild, node)
+- Add `build` workflow (esbuild, turbo)
 - Add `dev` workflow (php, livewire)